+++ src/main/java/ru/itgirl/libraryproject/service/BookServiceImpl.java
@@ -5,39 +5,57 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import ru.itgirl.libraryproject.dto.BookCreateDto;
-import ru.itgirl.libraryproject.dto.BookDto;
-import ru.itgirl.libraryproject.dto.BookUpdateDto;
-import ru.itgirl.libraryproject.dto.GenreCreateDto;
+import ru.itgirl.libraryproject.dto.*;
+import ru.itgirl.libraryproject.model.Author;
 import ru.itgirl.libraryproject.model.Book;
 import ru.itgirl.libraryproject.model.Genre;
 import ru.itgirl.libraryproject.repository.BookRepository;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class BookServiceImpl implements BookService {
 
     private final BookRepository bookRepository;
     private final GenreService genreService;
     @Override
     public BookDto getByNameV1(String name) {
-        Book book = bookRepository.findBookByName(name).orElseThrow();
-        return convertEntityToDto(book);
+        log.info("Try to find book by name {}", name);
+        Optional<Book> book = bookRepository.findBookByName(name);
+        if (book.isPresent()) {
+            BookDto bookDto = convertEntityToDto(book.get());
+            log.info("Book: {}", bookDto.toString());
+            return bookDto;
+        } else {
+            log.error("Book with name: {} not found", name);
+            throw new IllegalStateException("Книга не найдена");
+        }
     }
 
     @Override
     public BookDto getByNameV2(String name) {
-        Book book = bookRepository.findBookByNameBySql(name).orElseThrow();
-        return convertEntityToDto(book);
+        log.info("Try to find book by name {}", name);
+        Optional<Book> book = bookRepository.findBookByNameBySql(name);
+        if (book.isPresent()) {
+            BookDto bookDto = convertEntityToDto(book.get());
+            log.info("Book: {}", bookDto.toString());
+            return bookDto;
+        } else {
+            log.error("Book with name: {} not found", name);
+            throw new IllegalStateException("Книга не найдена");
+        }
     }
 
     @Override
     public BookDto getByNameV3(String name) {
+        log.info("Try to find book by name {}", name);
         Specification<Book> bookSpecification = Specification.where(new Specification<Book>() {
             @Override
             public Predicate toPredicate(Root<Book> root,
@@ -46,43 +64,65 @@ public class BookServiceImpl implements BookService {
                 return criteriaBuilder.equal(root.get("name"), name);
             }
         });
-        Book book = bookRepository.findOne(bookSpecification).orElseThrow();
-        return convertEntityToDto(book);
+        Optional<Book> book = bookRepository.findOne(bookSpecification);
+        if (book.isPresent()) {
+            BookDto bookDto = convertEntityToDto(book.get());
+            log.info("Book: {}", bookDto.toString());
+            return bookDto;
+        } else {
+            log.error("Book with name: {} not found", name);
+            throw new IllegalStateException("Книга не найдена");
+        }
     }
 
     @Override
     public BookDto createBook(BookCreateDto bookCreateDto) {
-        //1 Достать genre из bookCreateDto
-        //2 Получить его id из таблицы Genre
-        //3 Сформировать bookCreateDto где есть название книги и GenreDto
-        //4  и теперь можно сохранять
+        log.info("Try to create book {} {}", bookCreateDto.getName(), bookCreateDto.getGenre());
         GenreCreateDto genreCreateDto = genreService.getGenreByName(bookCreateDto.getGenre());
         Book book = bookRepository.save(convertDtoToEntity(bookCreateDto, genreCreateDto));
         BookDto bookDto = convertEntityToDto(book);
+        log.info("Book is created {} {}", book.getName(), book.getGenre());
         return bookDto;
     }
 
     @Override
     public BookDto updateBook(BookUpdateDto bookUpdateDto) {
-        Book book = bookRepository.findById(bookUpdateDto.getId()).orElseThrow();
+
+        log.info("Try to update book by id {}", bookUpdateDto.getId());
+        Optional<Book> book = bookRepository.findById(bookUpdateDto.getId());
         Genre genre = genreService.getGenreByName2(bookUpdateDto.getGenre());
-
-        book.setName(bookUpdateDto.getName());
-        book.setGenre(genre);
-
-        Book savedBook = bookRepository.save(book);
-        BookDto bookDto = convertEntityToDto(savedBook);
-        return bookDto;
+        if (book.isPresent()) {
+            Book bookObj = book.get();
+            bookObj.setName(bookUpdateDto.getName());
+            bookObj.setGenre(genre);
+            Book savedBook = bookRepository.save(bookObj);
+            BookDto bookDto = convertEntityToDto(savedBook);
+            log.info("Book was updated {} {}", bookDto.getName(), bookDto.getGenre());
+            return bookDto;
+        } else {
+            log.error("Book with id: {} not found", bookUpdateDto.getId());
+            throw new IllegalStateException("Книга не найдена");
+        }
     }
 
     @Override
     public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+        log.info("Try to delete book by id {}", id);
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()) {
+            bookRepository.deleteById(id);
+            log.info("Book was deleted, id = {}", id);
+        } else {
+            log.error("Book with id: {} not found", id);
+            throw new IllegalStateException("Книга не найден");
+        }
     }
 
     @Override
     public List<BookDto> getAllBooks() {
+        log.info("Try to get all books");
         List<Book> books = bookRepository.findAll();
+        log.info("All books got");
         return books.stream().map(this::convertEntityToDto).collect(Collectors.toList());
     }
